fix - добавлены тесты, добавлены интерфейсы
