+++ README.md
@@ -13,6 +13,7 @@ Template repository for Filmorate project.
 А также добавлены две новые фичи:
  - Функциональность «Популярные фильмы», которая предусматривает вывод самых любимых у зрителей фильмов по жанрам и годам.
  - Функциональность «Фильмы по режиссёрам», которая предполагает добавление к фильму информации о его режиссёре.
+ - Функциональность «Удаление фильмов и пользователей», которая предусматривает удаление фильма и пользователя по идентификатору.
 
 ![](Schema2.png)
