+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreStorage.java
@@ -1,38 +1,12 @@
 package ru.yandex.practicum.filmorate.storage.genre;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 
 import java.util.Set;
-import java.util.TreeSet;
 
-@Component
-public class GenreStorage {
-    JdbcTemplate jdbcTemplate;
-    public GenreStorage(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
+public interface GenreStorage {
+    Set<Genre> getFilmGenres(Long filmId);
 
-    public Set<Genre> getFilmGenres(Long filmId) {
-        String sql = "SELECT GENRES.GENRE_ID, GENRE FROM FILM_GENRES JOIN GENRES " +
-                "ON FILM_GENRES.GENRE_ID = GENRES.GENRE_ID " +
-                "WHERE FILM_ID = ?";
-        return new TreeSet<>(jdbcTemplate.query(sql, (rs, rowNum) -> new Genre(
-                        rs.getInt("genre_id"),
-                        rs.getString("genre")),
-                filmId
-        ));
-    }
-
-    public void updateGenresOfFilm(Film film) {
-        jdbcTemplate.update("DELETE FROM FILM_GENRES WHERE film_id = ?", film.getId());
-        if (film.getGenres() != null) {
-            for (Genre genre : film.getGenres()) {
-                jdbcTemplate.update("INSERT INTO FILM_GENRES (film_id, genre_id) VALUES (?, ?)",
-                        film.getId(), genre.getId());
-            }
-        }
-    }
+    void updateGenresOfFilm(Film film);
 }
