+++ src/test/java/ru/yandex/practicum/filmorate/UserControllerMockMvcTest.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -19,12 +19,11 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @SpringBootTest
 @AutoConfigureMockMvc
@@ -39,6 +38,14 @@ public class UserControllerMockMvcTest {
     private UserController controller;
     @Autowired
     private ObjectMapper objectMapper;
+
+    @BeforeEach
+    void clearTables() {
+        jdbcTemplate.update("DELETE FROM FILMS");
+        jdbcTemplate.update("DELETE FROM USERS");
+        jdbcTemplate.update("ALTER TABLE FILMS ALTER COLUMN FILM_ID RESTART WITH 1");
+        jdbcTemplate.update("ALTER TABLE USERS ALTER COLUMN USER_ID RESTART WITH 1");
+    }
 
     @Test
     void shouldAddUser() throws Exception {
@@ -138,6 +145,34 @@ public class UserControllerMockMvcTest {
     }
 
     @Test
+    void testFeed() throws Exception {
+        User userOne = new User(null, "mail@mail.ru", "dolore", "Nick Name",
+                LocalDate.of(1946, 8, 20));
+        User userTwo = new User(null, "mail@yandex.ru", "qwerty", "Nick Name",
+                LocalDate.of(1948, 7, 25));
+        mockMvc.perform(
+                post("/users")
+                        .content(objectMapper.writeValueAsString(userOne))
+                        .contentType(MediaType.APPLICATION_JSON)
+        ).andExpect(status().isOk());
+        mockMvc.perform(
+                post("/users")
+                        .content(objectMapper.writeValueAsString(userTwo))
+                        .contentType(MediaType.APPLICATION_JSON)
+        ).andExpect(status().isOk());
+        mockMvc.perform(
+                put("/users/1/friends/2")
+        ).andExpect(status().isOk());
+        mockMvc.perform(
+                get("/users/1/feed")
+        ).andExpect(status().isOk())
+                .andExpect(jsonPath("$.*", hasSize(1)))
+                .andExpect(jsonPath("$[0].eventType").value("FRIEND"))
+                .andExpect(jsonPath("$[0].operation").value("ADD"))
+                .andExpect(jsonPath("$[0].entityId").value("2"));
+    }
+
+    @Test
     void testCommonFilms() throws Exception {
         String sqlQuery = "DELETE FROM films";
         jdbcTemplate.update(sqlQuery);
@@ -159,13 +194,13 @@ public class UserControllerMockMvcTest {
                 200,
                 null,
                 new Mpa(1, "G"));
-    Film film3 = new Film(3L,
-            "New film2",
-            "Film description2",
-            LocalDate.of(2021, 2, 22),
-            200,
-            null,
-            new Mpa(1, "G"));
+        Film film3 = new Film(3L,
+                "New film2",
+                "Film description2",
+                LocalDate.of(2021, 2, 22),
+                200,
+                null,
+                new Mpa(1, "G"));
         User user1 = new User(1L,
                 "mail@mail.ru",
                 "Tommi",
@@ -186,11 +221,11 @@ public class UserControllerMockMvcTest {
                         .content(objectMapper.writeValueAsString(film2))
                         .contentType(MediaType.APPLICATION_JSON)
         ).andExpect(status().isOk());
-    mockMvc.perform(
-            post("/films")
-                    .content(objectMapper.writeValueAsString(film3))
-                    .contentType(MediaType.APPLICATION_JSON)
-    ).andExpect(status().isOk());
+        mockMvc.perform(
+                post("/films")
+                        .content(objectMapper.writeValueAsString(film3))
+                        .contentType(MediaType.APPLICATION_JSON)
+        ).andExpect(status().isOk());
         mockMvc.perform(
                 post("/users")
                         .content(objectMapper.writeValueAsString(user1))
@@ -211,11 +246,11 @@ public class UserControllerMockMvcTest {
         mockMvc.perform(put("/films/2/like/2"))
                 .andExpect(status().isOk());
         assertEquals(1, controller.getRecommendations(1L).size());
-        assertEquals("New film",new ArrayList<>(controller.getRecommendations(1L)).get(0).getName());
+        assertEquals("New film", new ArrayList<>(controller.getRecommendations(1L)).get(0).getName());
         mockMvc.perform(put("/films/3/like/2"))
                 .andExpect(status().isOk());
         assertEquals(2, controller.getRecommendations(1L).size());
-        assertEquals("New film",new ArrayList<>(controller.getRecommendations(1L)).get(0).getName());
-        assertEquals("New film2",new ArrayList<>(controller.getRecommendations(1L)).get(1).getName());
+        assertEquals("New film", new ArrayList<>(controller.getRecommendations(1L)).get(0).getName());
+        assertEquals("New film2", new ArrayList<>(controller.getRecommendations(1L)).get(1).getName());
     }
 }
